gmsol_treasury

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

cancel_swap

claim_fees

complete_gt_exchange

confirm_gt_buyback

create_swap

deposit_to_treasury_vault

initialize_config

initialize_treasury_vault_config

insert_token_to_treasury_vault

prepare_gt_bank

remove_token_from_treasury_vault

set_buyback_factor

set_gt_factor

set_referral_reward

set_treasury_vault_config

sync_gt_bank

toggle_token_flag

transfer_receiver

withdraw_from_treasury_vault