gmsol_treasury::cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

CancelSwap

Generated CPI struct of the accounts for CancelSwap.

ClaimFees

Generated CPI struct of the accounts for ClaimFees.

CompleteGtExchange

Generated CPI struct of the accounts for CompleteGtExchange.

ConfirmGtBuyback

Generated CPI struct of the accounts for ConfirmGtBuyback.

CreateSwap

Generated CPI struct of the accounts for CreateSwap.

DepositToTreasuryVault

Generated CPI struct of the accounts for DepositToTreasuryVault.

InitializeConfig

Generated CPI struct of the accounts for InitializeConfig.

InitializeTreasuryVaultConfig

Generated CPI struct of the accounts for InitializeTreasuryVaultConfig.

InsertTokenToTreasuryVault

Generated CPI struct of the accounts for InsertTokenToTreasuryVault.

PrepareGtBank

Generated CPI struct of the accounts for PrepareGtBank.

RemoveTokenFromTreasuryVault

Generated CPI struct of the accounts for RemoveTokenFromTreasuryVault.

SetReferralReward

Generated CPI struct of the accounts for SetReferralReward.

SetTreasuryVaultConfig

Generated CPI struct of the accounts for SetTreasuryVaultConfig.

SyncGtBank

Generated CPI struct of the accounts for SyncGtBank.

ToggleTokenFlag

Generated CPI struct of the accounts for ToggleTokenFlag.

TransferReceiver

Generated CPI struct of the accounts for TransferReceiver.

UpdateConfig

Generated CPI struct of the accounts for UpdateConfig.

WithdrawFromTreasuryVault

Generated CPI struct of the accounts for WithdrawFromTreasuryVault.